+++ package.json
@@ -13,6 +13,7 @@
   "dependencies": {
     "@react-oauth/google": "^0.10.0",
     "axios": "^1.3.5",
+    "jsonwebtoken": "^9.0.0",
     "jwt-decode": "^3.1.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
